refactoring installer folder
