+++ .env
@@ -40,8 +40,8 @@ CUSTOMER_CODE=
 TWILIO_PHONE_NUMBER=
 
 # format: text
-# required: true
-# configurable: true
+# required: false
+# configurable: false
 # description: Twilio Messaging Service SID;  It is used for sending scheduled messages
 MESSAGING_SERVICE_SID=
 
@@ -76,7 +76,7 @@ REMINDER_SECOND_TIMING=
 # description: Wait time for book or remind appointment message
 # format: text
 # required: true
-# configurable: false
-# default: 120
+# configurable: true
+# default: 5
 REPLY_WAIT_TIME=
